More analyzer adjustments

+++ src/Lingualizr.Tests/Localisation/es/NumberToWordsTests.cs
@@ -368,12 +368,12 @@ public class NumberToWordsTests
     }
 
     [Theory]
-    [InlineData((long)1, WordForm.Abbreviation, "un")]
-    [InlineData((long)1, WordForm.Normal, "uno")]
-    [InlineData((long)21, WordForm.Abbreviation, "veintiún")]
-    [InlineData((long)21, WordForm.Normal, "veintiuno")]
-    [InlineData((long)21501, WordForm.Abbreviation, "veintiún mil quinientos un")]
-    [InlineData((long)21501, WordForm.Normal, "veintiún mil quinientos uno")]
+    [InlineData(1L, WordForm.Abbreviation, "un")]
+    [InlineData(1L, WordForm.Normal, "uno")]
+    [InlineData(21L, WordForm.Abbreviation, "veintiún")]
+    [InlineData(21L, WordForm.Normal, "veintiuno")]
+    [InlineData(21501L, WordForm.Abbreviation, "veintiún mil quinientos un")]
+    [InlineData(21501L, WordForm.Normal, "veintiún mil quinientos uno")]
     public void ToWordsLongWithWordForm(long number, WordForm wordForm, string expected)
     {
         Assert.Equal(expected, number.ToWords(wordForm));
@@ -382,13 +382,13 @@ public class NumberToWordsTests
     }
 
     [Theory]
-    [InlineData((long)1, WordForm.Abbreviation, GrammaticalGender.Masculine, "un")]
-    [InlineData((long)1, WordForm.Abbreviation, GrammaticalGender.Feminine, "una")]
-    [InlineData((long)21, WordForm.Abbreviation, GrammaticalGender.Masculine, "veintiún")]
-    [InlineData((long)21, WordForm.Abbreviation, GrammaticalGender.Feminine, "veintiuna")]
-    [InlineData((long)21501, WordForm.Abbreviation, GrammaticalGender.Masculine, "veintiún mil quinientos un")]
-    [InlineData((long)21501, WordForm.Normal, GrammaticalGender.Masculine, "veintiún mil quinientos uno")]
-    [InlineData((long)21501, WordForm.Abbreviation, GrammaticalGender.Feminine, "veintiuna mil quinientas una")]
+    [InlineData(1L, WordForm.Abbreviation, GrammaticalGender.Masculine, "un")]
+    [InlineData(1L, WordForm.Abbreviation, GrammaticalGender.Feminine, "una")]
+    [InlineData(21L, WordForm.Abbreviation, GrammaticalGender.Masculine, "veintiún")]
+    [InlineData(21L, WordForm.Abbreviation, GrammaticalGender.Feminine, "veintiuna")]
+    [InlineData(21501L, WordForm.Abbreviation, GrammaticalGender.Masculine, "veintiún mil quinientos un")]
+    [InlineData(21501L, WordForm.Normal, GrammaticalGender.Masculine, "veintiún mil quinientos uno")]
+    [InlineData(21501L, WordForm.Abbreviation, GrammaticalGender.Feminine, "veintiuna mil quinientas una")]
     public void ToWordsLongWithWordFormAndGender(long number, WordForm wordForm, GrammaticalGender gender, string expected)
     {
         Assert.Equal(expected, number.ToWords(wordForm, gender));
